+++ internal/models/request/bulk.go
