--- lambda/service/go.mod
@@ -1,16 +1,0 @@
-module github.com/pennsieve/drs-service/service
-
-//TODO update module name
-
-go 1.21
-
-require (
-	github.com/aws/aws-lambda-go v1.34.1
-	github.com/stretchr/testify v1.7.2
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
