--- local-drs-service/go.mod
@@ -1,14 +1,0 @@
-module github.com/daniellong/local-drs-service
-
-go 1.23.4
-
-require (
-	github.com/gorilla/mux v1.8.1
-	github.com/stretchr/testify v1.10.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
